rdcrdcrdc

+++ src/main/java/com/dit/hua/houseM/repositories/AdminRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdminRepository extends JpaRepository<Admin, Long> {
+public interface AdminRepository extends JpaRepository<Admin, Integer> {
 
     // Custom query to find an admin by username
     Admin findByUsername(String username);
