+++ src/main/java/com/dit/hua/houseM/repositories/OwnerRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface OwnerRepository extends JpaRepository<Owner, Long> {
+public interface OwnerRepository extends JpaRepository<Owner, Integer> {
 
     Optional<Owner> findByUsername(String username);
